Move AlluxioWorker behind an interface
